Try to add all available certificates from the Firefox Certificate Store.
We iterate all certs from the Firefox Cert Store, and then we perform a 
getChain() for each one of them. We make sure we do not add duplicates.
We might revise this in the future when we get a better understanding
on which certs make it to the Certificate Store.