Respect preferences when showing certificates.
We check each setting before we decide to show the certificate dialog.
For each certificate in the chain we show a star if a preference
matches for this certificate. For example, if we show website certs
on first time only, we add a star on the tab if indeed the website cert
if found for the first time.